pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindweave"
version = "0.1.0"
description = "MindWeave AI Environment"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    # Core API
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",

    # ML stack
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "sentence-transformers>=2.6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",

    # OpenEnv
    "openenv>=0.1.13",
    "openenv-core[core]>=0.2.2",
]

[project.entry-points."openenv.envs"]
mindweave = "server.environment2:MindweaveEnvironment"

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["server*"]
exclude = ["logs*", "models*"]